move counter to testutils
