+++ testutil/x/counter/proto/buf.gen.gogo.yaml
