+++ testutil/x/counter/proto/buf.lock
