+++ testutil/x/counter/proto/cosmos/counter/module/v1/module.proto
@@ -7,7 +7,7 @@ import "cosmos/app/v1alpha1/module.proto";
 // Module is the config object of the counter module.
 message Module {
   option (cosmos.app.v1alpha1.module) = {
-    go_import: "github.com/cosmos/cosmos-sdk/x/counter"
+    go_import: "github.com/cosmos/cosmos-sdk/testutil/x/counter"
   };
 
   // authority defines the custom module authority. If not set, defaults to the governance module.
