+++ testutil/x/counter/proto/cosmos/counter/v1/query.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 package cosmos.counter.v1;
 
-option go_package = "github.com/cosmos/cosmos-sdk/x/counter/types";
+option go_package = "github.com/cosmos/cosmos-sdk/testutil/x/counter/types";
 
 // Query defines the gRPC querier service.
 service Query {
