+++ testutil/x/counter/proto/cosmos/counter/v1/tx.proto
@@ -6,7 +6,7 @@ import "cosmos_proto/cosmos.proto";
 import "cosmos/msg/v1/msg.proto";
 import "amino/amino.proto";
 
-option go_package = "github.com/cosmos/cosmos-sdk/x/counter/types";
+option go_package = "github.com/cosmos/cosmos-sdk/testutil/x/counter/types";
 
 // Msg defines the counter Msg service.
 service Msg {
